refactor: decodeFlags/encodeFlags/타입 중복 제거
- Zig: decodeFlags()를 src/transpile.zig로 추출, wasm/ffi 진입점에서 공유
- TypeScript: 타입(TranspileOptions 등), encodeFlags(), ES_TARGET_BITS를
  packages/shared/index.ts로 추출, core/wasm에서 import
