+++ src/transpile.zig
@@ -63,6 +63,52 @@ pub const TranspileOptions = struct {
     es_target: ?@import("transformer/compat.zig").ESTarget = null,
 };
 
+/// WASM/FFI 진입점 공용 플래그 디코더.
+/// 비트마스크 → TranspileOptions 변환 (비트 레이아웃은 packages/wasm, packages/core 참조).
+pub fn decodeFlags(flags: u32) TranspileOptions {
+    return .{
+        .sourcemap = flags & (1 << 0) != 0,
+        .minify_whitespace = flags & (1 << 1) != 0,
+        .minify_identifiers = flags & (1 << 2) != 0,
+        .minify_syntax = flags & (1 << 3) != 0,
+        .jsx_runtime = if (flags & (1 << 5) != 0)
+            .automatic_dev
+        else if (flags & (1 << 4) != 0)
+            .automatic
+        else
+            .classic,
+        .drop_console = flags & (1 << 6) != 0,
+        .drop_debugger = flags & (1 << 7) != 0,
+        .ascii_only = flags & (1 << 8) != 0,
+        .flow = flags & (1 << 9) != 0,
+        .experimental_decorators = flags & (1 << 10) != 0,
+        .emit_decorator_metadata = flags & (1 << 11) != 0,
+        .module_format = switch ((flags >> 12) & 0x3) {
+            0 => .esm,
+            1 => .cjs,
+            else => .esm,
+        },
+        .quote_style = switch ((flags >> 14) & 0x3) {
+            0 => .double,
+            1 => .single,
+            2 => .preserve,
+            else => .double,
+        },
+        .use_define_for_class_fields = flags & (1 << 16) != 0,
+        .charset_utf8 = flags & (1 << 17) != 0,
+        .platform = switch ((flags >> 18) & 0x3) {
+            0 => .browser,
+            1 => .node,
+            2 => .neutral,
+            3 => .react_native,
+            else => .browser,
+        },
+        .jsx_in_js = flags & (1 << 20) != 0,
+        .sourcemap_debug_ids = flags & (1 << 21) != 0,
+        .sources_content = flags & (1 << 22) != 0,
+    };
+}
+
 pub const TranspileError = error{
     ParseError,
     SemanticError,
